requirements.txt
websocket-client
pillow
